Human prompts with markdown styling should be styled Fixes #117
